move initializeFirebaseAdmin to a new file
