+++ src/fixtures/firebase/index.ts
@@ -1,7 +1,5 @@
 import { initializeApp } from 'firebase/app'
 import { getAuth } from 'firebase/auth'
-import { initializeApp as initializeAdminApp } from 'firebase-admin/app'
-import { getAuth as getAdminAuth } from 'firebase-admin/auth'
 
 export const initializeFirebase = () => {
   // Your web app's Firebase configuration
@@ -21,15 +19,4 @@ export const initializeFirebase = () => {
   return auth
 }
 
-export const initializeFirebaseAdmin = () => {
-  // Your web app's Firebase configuration.
-  const firebaseConfig = {
-    projectId: import.meta.env.PUBLIC_FIREBASE_PROJECT_ID,
-  }
 
-  // Initialize Firebase Admin.
-  const app = initializeAdminApp(firebaseConfig)
-  const auth = getAdminAuth(app)
-
-  return auth
-}
